Converted assigned components to tsx ext; checked that application runs correctly on both client and server sides with current code
